requirements.txt
fastapi
uvicorn
sse-starlette
requests
sseclient-py